chore: move btc vault swap functions out of test and rename test
