+++ bouncer/shared/perform_swap.ts
@@ -30,7 +30,7 @@ import { SwapContext, SwapStatus } from './swap_context';
 import { getChainflipApi, observeEvent } from './utils/substrate';
 import { executeEvmVaultSwap } from './evm_vault_swap';
 import { executeSolVaultSwap } from './sol_vault_swap';
-import { buildAndSendBtcVaultSwap, openPrivateBtcChannel } from '../tests/btc_vault_swap';
+import { buildAndSendBtcVaultSwap, openPrivateBtcChannel } from './btc_vault_swap';
 
 function encodeDestinationAddress(address: string, destAsset: Asset): string {
   let destAddress = address;
