--- bouncer/tests/btc_vault_swap.ts
@@ -1,215 +1,0 @@
-import assert from 'assert';
-import { Chains } from '@chainflip/cli';
-import { ExecutableTest } from '../shared/executable_test';
-import { BTC_ENDPOINT, waitForBtcTransaction, sendVaultTransaction } from '../shared/send_btc';
-import {
-  amountToFineAmount,
-  Asset,
-  assetDecimals,
-  brokerMutex,
-  chainFromAsset,
-  createStateChainKeypair,
-  decodeDotAddressForContract,
-  fineAmountToAmount,
-  handleSubstrateError,
-  newAddress,
-  observeBalanceIncrease,
-  stateChainAssetFromAsset,
-} from '../shared/utils';
-import { getChainflipApi, observeEvent } from '../shared/utils/substrate';
-import { getBalance } from '../shared/get_balance';
-import { getEarnedBrokerFees } from './broker_fee_collection';
-import { fundFlip } from '../shared/fund_flip';
-
-/* eslint-disable @typescript-eslint/no-use-before-define */
-export const testBtcVaultSwap = new ExecutableTest('Btc-Vault-Swap', main, 120);
-
-// Fee to use for the broker and affiliates
-const commissionBps = 100;
-
-interface BtcVaultSwapDetails {
-  chain: string;
-  nulldata_payload: string;
-  deposit_address: string;
-}
-
-interface BtcVaultSwapExtraParameters {
-  chain: 'Bitcoin';
-  min_output_amount: string;
-  retry_duration: number;
-}
-
-interface Beneficiary {
-  account: string;
-  bps: number;
-}
-
-export async function buildAndSendBtcVaultSwap(
-  depositAmountBtc: number,
-  brokerUri: string,
-  destinationAsset: Asset,
-  destinationAddress: string,
-  refundAddress: string,
-  affiliateAddresses: string[],
-) {
-  await using chainflip = await getChainflipApi();
-
-  const broker = createStateChainKeypair(brokerUri);
-  testBtcVaultSwap.debugLog('Broker:', broker.address);
-  testBtcVaultSwap.debugLog(`Btc endpoint is set to`, BTC_ENDPOINT);
-
-  const affiliates: Beneficiary[] = [];
-  for (const affiliateAddress of affiliateAddresses) {
-    affiliates.push({ account: affiliateAddress, bps: commissionBps });
-  }
-
-  const extraParameters: BtcVaultSwapExtraParameters = {
-    chain: 'Bitcoin',
-    min_output_amount: '0',
-    retry_duration: 0,
-  };
-
-  const BtcVaultSwapDetails = (await chainflip.rpc(
-    `cf_get_vault_swap_details`,
-    broker.address,
-    { chain: 'Bitcoin', asset: stateChainAssetFromAsset('Btc') },
-    { chain: chainFromAsset(destinationAsset), asset: stateChainAssetFromAsset(destinationAsset) },
-    chainFromAsset(destinationAsset) === Chains.Polkadot
-      ? decodeDotAddressForContract(destinationAddress)
-      : destinationAddress,
-    commissionBps, // broker_commission
-    extraParameters,
-    null, // channel_metadata
-    0, // boost_fee
-    affiliates,
-    null, // dca_params
-  )) as unknown as BtcVaultSwapDetails;
-
-  assert.strictEqual(BtcVaultSwapDetails.chain, 'Bitcoin');
-  testBtcVaultSwap.debugLog('nulldata_payload:', BtcVaultSwapDetails.nulldata_payload);
-  testBtcVaultSwap.debugLog('deposit_address:', BtcVaultSwapDetails.deposit_address);
-
-  const txid = await sendVaultTransaction(
-    BtcVaultSwapDetails.nulldata_payload,
-    depositAmountBtc,
-    BtcVaultSwapDetails.deposit_address,
-    refundAddress,
-  );
-  testBtcVaultSwap.log('Broadcast successful, txid:', txid);
-
-  await waitForBtcTransaction(txid);
-  testBtcVaultSwap.debugLog('Transaction confirmed');
-  return txid;
-}
-
-async function testVaultSwap(
-  depositAmountBtc: number,
-  brokerUri: string,
-  destinationAsset: Asset,
-  affiliateUri: string,
-) {
-  // Addresses
-  const destinationAddress = await newAddress(destinationAsset, 'BTC_VAULT_SWAP');
-  testBtcVaultSwap.debugLog('destinationAddress:', destinationAddress);
-  const refundAddress = await newAddress('Btc', 'BTC_VAULT_SWAP_REFUND');
-  testBtcVaultSwap.debugLog('Refund address:', refundAddress);
-
-  // Amounts before swap
-  const destinationAmountBeforeSwap = await getBalance(destinationAsset, destinationAddress);
-  const broker = createStateChainKeypair(brokerUri);
-  const affiliate = createStateChainKeypair(affiliateUri);
-  const earnedBrokerFeesBefore = await getEarnedBrokerFees(broker);
-  const earnedAffiliateFeesBefore = await getEarnedBrokerFees(affiliate);
-  testBtcVaultSwap.debugLog('Earned broker fees before:', earnedBrokerFeesBefore);
-  testBtcVaultSwap.debugLog('Earned affiliate fees before:', earnedAffiliateFeesBefore);
-
-  const observeSwapExecutedEvent = observeEvent(`swapping:SwapExecuted`, {
-    test: (event) =>
-      event.data.inputAsset === 'Btc' &&
-      event.data.outputAsset === destinationAsset &&
-      event.data.inputAmount.replace(/,/g, '') ===
-        amountToFineAmount(depositAmountBtc.toString(), assetDecimals('Btc')),
-  }).event;
-
-  await buildAndSendBtcVaultSwap(
-    depositAmountBtc,
-    brokerUri,
-    destinationAsset,
-    destinationAddress,
-    refundAddress,
-    [affiliate.address],
-  );
-
-  // Complete swap and check balance
-  testBtcVaultSwap.debugLog('Waiting for swap executed event');
-  await observeSwapExecutedEvent;
-  testBtcVaultSwap.log(`Btc -> ${destinationAsset} Vault Swap executed`);
-  await observeBalanceIncrease(destinationAsset, destinationAddress, destinationAmountBeforeSwap);
-  testBtcVaultSwap.log(`Balance increased, Vault Swap Complete`);
-
-  // Check that both the broker and affiliate earned fees
-  const earnedBrokerFeesAfter = await getEarnedBrokerFees(broker);
-  const earnedAffiliateFeesAfter = await getEarnedBrokerFees(affiliate);
-  testBtcVaultSwap.debugLog('Earned broker fees after:', earnedBrokerFeesAfter);
-  testBtcVaultSwap.debugLog('Earned affiliate fees after:', earnedAffiliateFeesAfter);
-  assert(earnedBrokerFeesAfter > earnedBrokerFeesBefore, 'No increase in earned broker fees');
-  assert(
-    earnedAffiliateFeesAfter > earnedAffiliateFeesBefore,
-    'No increase in earned affiliate fees',
-  );
-}
-
-export async function openPrivateBtcChannel(brokerUri: string) {
-  // Check if the channel is already open
-  const chainflip = await getChainflipApi();
-  const broker = createStateChainKeypair(brokerUri);
-  const existingPrivateChannel = Number(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (await chainflip.query.swapping.brokerPrivateBtcChannels(broker.address)) as any,
-  );
-
-  if (!existingPrivateChannel) {
-    // Fund the broker the required bond amount for opening a private channel
-    const fundAmount = fineAmountToAmount(
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (await chainflip.query.swapping.brokerBond()) as any as string,
-      assetDecimals('Flip'),
-    );
-    await fundFlip(broker.address, fundAmount);
-
-    // Open the private channel
-    await brokerMutex.runExclusive(async () => {
-      await chainflip.tx.swapping
-        .openPrivateBtcChannel()
-        .signAndSend(broker, { nonce: -1 }, handleSubstrateError(chainflip));
-    });
-    console.log('Private Btc channel opened');
-  }
-}
-
-async function registerAffiliate(
-  brokerUri: string,
-  affiliateUri: string,
-  affiliateShortId: number,
-) {
-  const chainflip = await getChainflipApi();
-  const broker = createStateChainKeypair(brokerUri);
-  const affiliate = createStateChainKeypair(affiliateUri);
-
-  await brokerMutex.runExclusive(async () => {
-    await chainflip.tx.swapping
-      .registerAffiliate(affiliate.address, affiliateShortId)
-      .signAndSend(broker, { nonce: -1 }, handleSubstrateError(chainflip));
-  });
-}
-
-async function main() {
-  const btcDepositAmount = 0.1;
-  // TODO: Fee collection will work properly when using 'BROKER_1' and 'BROKER_2' because it will be effected by the other tests.
-  const brokerUri = '//BROKER_1';
-  const affiliateUri = '//BROKER_2';
-
-  await openPrivateBtcChannel(brokerUri);
-  await registerAffiliate(brokerUri, affiliateUri, 0);
-  await testVaultSwap(btcDepositAmount, brokerUri, 'Flip', affiliateUri);
-}
